+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * Support for creating AcroForm password fields and their appearances
 * Support for creating AcroForm file select fields and their appearances
 
+### Fixed
+
+* Missing usage of `:sort` flag for AcroForm choice fields
+
 
 ## 0.13.0 - 2020-11-15